Move VGradeFormatter tests to dedicated test file
Extract VGradeFormatter tests from SessionWebSocketManagerTests.swift
into a new VGradeFormatterTests.swift for better discoverability.
No test logic changes.
